+++ src/mantle/http-client/class-pooled-pending-request.php
@@ -37,10 +37,10 @@ class Pooled_Pending_Request extends Pending_Request {
 	/**
 	 * Issue a GET request to the given URL.
 	 *
-	 * @param  string                           $url URL to retrieve.
+	 * @param  string|null                      $url URL to retrieve.
 	 * @param  array<string, mixed>|string|null $query Query parameters (assumed to be urlencoded).
 	 */
-	public function get( string $url, array|string|null $query = null ): static {
+	public function get( ?string $url = null, array|string|null $query = null ): static {
 		$this->set_method( Http_Method::GET )->set_url( $url );
 
 		if ( $query ) {
@@ -53,10 +53,10 @@ class Pooled_Pending_Request extends Pending_Request {
 	/**
 	 * Issue a HEAD request to the given URL.
 	 *
-	 * @param  string                           $url URL to retrieve.
+	 * @param  string|null                      $url URL to retrieve.
 	 * @param  array<string, mixed>|string|null $query Query parameters (assumed to be urlencoded).
 	 */
-	public function head( string $url, array|string|null $query = null ): static {
+	public function head( ?string $url = null, array|string|null $query = null ): static {
 		$this->set_method( Http_Method::HEAD )->set_url( $url );
 
 		if ( $query ) {
@@ -69,10 +69,10 @@ class Pooled_Pending_Request extends Pending_Request {
 	/**
 	 * Issue a POST request to the given URL.
 	 *
-	 * @param  string                    $url URL to post.
+	 * @param  string|null               $url URL to post.
 	 * @param  array<string, mixed>|null $data Data to send with the request.
 	 */
-	public function post( string $url, ?array $data = null ): static {
+	public function post( ?string $url = null, ?array $data = null ): static {
 		$this->set_method( Http_Method::POST )->set_url( $url );
 
 		if ( $data ) {
@@ -85,10 +85,10 @@ class Pooled_Pending_Request extends Pending_Request {
 	/**
 	 * Issue a PATCH request to the given URL.
 	 *
-	 * @param  string                    $url URL to patch.
+	 * @param  string|null               $url URL to patch.
 	 * @param  array<string, mixed>|null $data Data to send with the request.
 	 */
-	public function patch( string $url, ?array $data = null ): static {
+	public function patch( ?string $url = null, ?array $data = null ): static {
 		$this->set_method( Http_Method::PATCH )->set_url( $url );
 
 		if ( $data ) {
@@ -101,10 +101,10 @@ class Pooled_Pending_Request extends Pending_Request {
 	/**
 	 * Issue a PUT request to the given URL.
 	 *
-	 * @param  string                    $url URL to put.
+	 * @param  string|null               $url URL to put.
 	 * @param  array<string, mixed>|null $data Data to send with the request.
 	 */
-	public function put( string $url, ?array $data = null ): static {
+	public function put( ?string $url = null, ?array $data = null ): static {
 		$this->set_method( Http_Method::PUT )->set_url( $url );
 
 		if ( $data ) {
@@ -117,10 +117,10 @@ class Pooled_Pending_Request extends Pending_Request {
 	/**
 	 * Issue a DELETE request to the given URL.
 	 *
-	 * @param  string                    $url URL to delete.
+	 * @param  string|null               $url URL to delete.
 	 * @param  array<string, mixed>|null $data Data to send with the request.
 	 */
-	public function delete( string $url, ?array $data = null ): static {
+	public function delete( ?string $url = null, ?array $data = null ): static {
 		$this->set_method( Http_Method::DELETE )->set_url( $url );
 
 		if ( $data ) {